Fix compatibility with Redux store
There was a problem when using Redux in your app along with ExNavigation in which ex-nav's `store` would override your app's store, when using `Provider` and `connect`. This commit introduces a wrapper around `connect` that's used in Ex Nav, in order to pass "navigationStore" through the context, rather than "store".

This allows the use of Redux in your app and Ex-Navigation side-by-side, while still allowing the ability to integrate ex-nav with your app's redux store (for more debuggability and integration with app logic).

fbshipit-source-id: 20e966c