Add renderScene prop to StackNavigation
The background color in [ExNavigationStackItem](https://github.com/exponentjs/ex-navigation/blob/68cbb5b082f1ef5e4599686f9168132d8ede7737/src/ExNavigationStackItem.js#L85) was causing some flashes of white in my dark colored application. This was *mostly* fixed by setting a background color using the  property in , but a light vertical line was still shown when transitioning between screens.

This PR adds a  prop to  component that allows overriding props passed to , including the  prop that can be used to set a different background color for .

 (ExNavigationStackItem) is also now exported in the public API.

An example of usage:

Closes exponentjs/ex-navigation#302

fbshipit-source-id: c81895f