Use measure instead of measureInWindow to avoid position bug with translucent status bar on Android
There'\''s a bug with  in RN with translucent statusbar because it always adds the statusbar height. This uses  instead with does pretty much the same thing but doesn'\''t have that bug (it'\''s actually the proper way to do it since it measures relative to the root view).

Fixes #320

Closes exponent/ex-navigation#321

fbshipit-source-id: 5a4d554