New Style for Horizontal with NoAnimation in NavBar
This came from a need to not have the navbar flash when screens were transitioning, due to the navbar being rather static (think Facebook'\''s search bar).

It'\''s possible this is a config or param setting I over-looked ... and if others want this, could see expanding this to allow navbar animation to be set separately or toggled off independently from the screen transition. In the spirit of keeping it simple, this simply creates a new style one can pass in that does not set any transition/fade/animation for the navbar while allowing the rest of the screen to animate in the traditional  fashion.

Closes exponent/ex-navigation#243

fbshipit-source-id: 74791b8